+++ opentelemetry-prometheus/src/config.rs
@@ -1,7 +1,6 @@
 use core::fmt;
 use once_cell::sync::OnceCell;
-use opentelemetry::metrics::{MetricsError, Result};
-use opentelemetry_sdk::metrics::ManualReaderBuilder;
+use opentelemetry_sdk::metrics::{ManualReaderBuilder, MetricError, MetricResult};
 use std::sync::{Arc, Mutex};
 
 use crate::{Collector, PrometheusExporter, ResourceSelector};
@@ -116,7 +115,7 @@ impl ExporterBuilder {
     }
 
     /// Creates a new [PrometheusExporter] from this configuration.
-    pub fn build(self) -> Result<PrometheusExporter> {
+    pub fn build(self) -> MetricResult<PrometheusExporter> {
         let reader = Arc::new(self.reader.build());
 
         let collector = Collector {
@@ -135,7 +134,7 @@ impl ExporterBuilder {
         let registry = self.registry.unwrap_or_default();
         registry
             .register(Box::new(collector))
-            .map_err(|e| MetricsError::Other(e.to_string()))?;
+            .map_err(|e| MetricError::Other(e.to_string()))?;
 
         Ok(PrometheusExporter { reader })
     }